restructuring and fixing demo tests
