--- integrationtests/demo-test/tests.launch
@@ -1,4 +1,0 @@
-<launch>
-	<include file="$(find px4)/launch/gazebo_iris_empty_world.launch" />
-	<test test-name="test1" pkg="px4" type="test1.py" />
-</launch>
