+++ integrationtests/demo-tests/demo_tests.launch
@@ -1,0 +1,5 @@
+<launch>
+	<include file="$(find px4)/launch/gazebo_iris_empty_world.launch" />
+	<test test-name="arm" pkg="px4" type="arm_test.py" />
+	<test test-name="posctl" pkg="px4" type="posctl_test.py" />
+</launch>
